// src/types/data.types.ts
export type DataFieldType = 'integer' | 'float' | 'string' | 'boolean';

export type NestedDataMap = {
  [key: string]: DataFieldType | NestedDataMap;
};


export type MetricMeta = {
  type: DataFieldType;
  label?: string;
  unit?: string;
};

export interface ProfileDataPoints {
  [profileType: string]: NestedDataMap;
}

export const trackableProfileData: ProfileDataPoints = {
  general: {
    Profile: {
      numberOfProfileCategories: 'integer',
      profileCompletionPercentage: 'integer',
      numberOfBadges: 'integer',
      numberOfMilestones: 'integer',
      numberOfAchievements: 'integer',
      profileViews: 'integer',
      engagement: {
        numberOfLikes: 'integer',
        numberOfShares: 'integer',
        numberOfComments: 'integer',
        numberOfPosts: 'integer',
        numberOfGroupsJoined: 'integer',
        numberOfEventsAttended: 'integer',
        numberOfMessagesSent: 'integer',
        numberOfMessagesReceived: 'integer',
        numberOfFollowing: 'integer',
      },
      mypts: {
        pointsEarned: 'integer',
        pointsSpent: 'integer',
        pointsRedeemed: 'integer',
        pointsBalance: 'integer',
      },
      LinkedProducts: {
        numberOfProducts: 'integer',
        numberOfActiveProducts: 'integer',
        numberOfSoldProducts: 'integer',
        numberOfReturnedProducts: 'integer',
        numberOfReviewedProducts: 'integer',
      },
      Circles: {
        numberOfConnections: 'integer',
        numberOfAffiliations: 'integer',
        numberOfPendingConnections: 'integer',
        numberOfPendingAffiliations: 'integer',
        numberOfPendingInvitations: 'integer',
        numberOfContacts: 'integer',
      }
    },
    Plans: {
      numberOfPlans: 'integer',
      numberOfActivePlans: 'integer',
      numberOfEvents: 'integer',
      numberOfTasks: 'integer',
      numberOfReminders: 'integer',
      numberOfNotes: 'integer',
      numberOfActiveTasters: 'integer',
      numberOfPendingEvents: 'integer',
      numberOfCancelledEvents: 'integer',
      numberOfCompletedEvents: 'integer',
      numberOfPastEvents: 'integer',
    },
    Vault: {
      numberOfDocuments: 'integer',
      numberOfImages: 'integer',
      numberOfVideos: 'integer',
      numberOfAudioFiles: 'integer',
      numberOfFiles: 'integer',
      amountOfStorageUsed: 'float',
      amountOfStorageAvailable: 'float',
    },
    interaction: {
  totalInteractions: 'integer',
  upcomingInteractions: 'integer',
  completedInteractions: 'integer',
  cancelledInteractions: 'integer',
  autoGeneratedInteractions: 'integer',
  personalInteractions: 'integer',
  workInteractions: 'integer',
  businessInteractions: 'integer',
  networkingInteractions: 'integer',
  inPersonInteractions: 'integer',
  virtualInteractions: 'integer',
  interactionsWithRewards: 'integer',
  interactionsWithReminders: 'integer',
  interactionsWithAttachments: 'integer',
  interactionsWithLocation: 'integer',
  interactionsWithContextualEntity: 'integer',
  mostUsedInteractionMode: 'string',
  uniqueInteractionPartners: 'integer',
  averageDaysBetweenContacts: 'float',
  lastInteractionDate: 'string',
  nextScheduledInteractionDate: 'string'
}

  },
// ...
};